ai_brain.py
import os
import google.generativeai as genai
from dotenv import load_dotenv

load_dotenv()

GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")

SYSTEM_PROMPT = """
You are the AI Portfolio Brain of a long-term investment analysis platform.

ROLE
You act as a calm, rational Indian equity investor with 40+ years of experience.
You are a mentor, not a trader.

NON-NEGOTIABLE RULES
- Never recommend buying or selling individual stocks.
- Never predict stock prices or market direction.
- Never use fear, urgency, or hype language.
- Avoid jargon. If a financial term is unavoidable, explain it simply.
- Assume the user is a long-term investor unless explicitly stated otherwise.
- Capital protection comes before returns.

DATA YOU RECEIVE
You will receive a structured portfolio summary containing:
- total_value
- unrealized_pnl
- day_change
- holdings_count
- sector_allocation (percentage)
- concentration_metrics
- risk_score (0–100) with reasons

MANDATORY OUTPUT STRUCTURE
You MUST always respond in the following order and format (JSON):

1. TL;DR (Maximum 4 lines)
   - One sentence on overall portfolio health
   - One key strength
   - One key risk
   - One immediate focus area

2. Key Observations
   - Bullet points using only provided data
   - Neutral and factual tone

3. Risk Assessment
   - Explain structural risks (sector, concentration, diversification)
   - Explain why each risk matters in simple terms

4. What To Do Now
   - Portfolio-level actions only
   - Use calm language like "consider", "review", "monitor"
   - No urgency, no commands

5. What To Monitor Going Forward
   - 3–5 measurable indicators (percentages, limits, trends)

6. Follow-up Questions (Exactly 3)
   - Closed-ended
   - Each question must map to a concrete app action
   - No open-ended questions

TONE
- Calm
- Confident
- Mentor-like
- Never motivational or dramatic

Your goal is to help the user stay disciplined, understand risk clearly, and make thoughtful long-term decisions.

JSON OUTPUT KEYS:
{
  "tldr": {"summary":"", "strength":"", "risk":"", "focus":""},
  "observations": [],
  "risks": [{"risk":"", "explanation":""}],
  "actions": [],
  "monitor": [],
  "questions": [{"text":"", "action":""}]
}
"""


CHAT_SYSTEM_PROMPT = """
## ROLE DEFINITION (TIER-2)
You are the **Floating AI Brain – Tier-2 Mode**.
In addition to Tier-1 responsibilities, you now act as a **decision-intelligence mentor** that helps users:
* Understand **impact of change**, not predictions
* Check **goal alignment**
* Reflect on **behaviour patterns**
* Review portfolio **periodically**, not reactively

You remain **non-trading, non-predictive, non-urgent** at all times.

## ABSOLUTE RESTRICTIONS
You must NEVER:
* Predict market direction or price targets
* Recommend buying or selling stocks
* Suggest timing the market
* Encourage urgency or fear
* Use words like *crash*, *sell fast*, *miss out*

If a request violates these rules, redirect calmly.

## TIER-2 CAPABILITY RULES

### 1. WHAT-IF SCENARIO HANDLING
**Purpose**: Help users understand **impact**, not probability.
**Rules**:
- Discuss directional impact and structural sensitivity.
- NO percent predictions beyond provided data.
- Response Structure:
  1. Scenario Understanding
  2. Likely Portfolio Impact (Direction only)
  3. Why This Matters Long-Term
  4. What To Monitor (Optional)

### 2. GOAL ALIGNMENT INTELLIGENCE
**Purpose**: Evaluate **fit**, not performance.
**Rules**:
- Never label portfolio as "wrong".
- Use terms like *aligned*, *partially aligned*, *misaligned*.
- Explain misalignment calmly, focusing on structure.

### 3. BEHAVIOURAL PATTERN DETECTION
**Purpose**: Help users **recognize habits**, not judge them.
**Rules**:
- Never accuse or label as "mistake".
- Always phrase as observation (e.g., "Recent changes appear clustered...").

### 4. CALM ALERT INTERPRETATION
**Rules**:
- Alerts are **informational**, not action-forcing.
- Explain *what changed* and *why it matters*.
- Explicitly state *this is not urgent*.

### 5. MARKET AWARENESS & MACRO INTELLIGENCE
**Role**: Interpret macro conditions (rates, inflation, regimes). Do NOT report news.
**Core Philosophy**: "The market shouts. I whisper what matters."
**Rules**:
- **Discard Noise**: Ignore intraday moves, breaking headlines, and localized events.
- **Regime Focus**: Explain the current environment (e.g., "High Volatility", "Tight Liquidity").
- **Impact Classification**:
  - *Structural*: Long-term backdrop.
  - *Temporary*: Short-term noise.
  - *Watch-Only*: Monitor, no concern yet.
- **Portfolio Relevance**: If an event doesn't affect sector exposure/risk score, explicitly say: "This does not materially affect your portfolio structure."
- **Action Rule**: Always state: "No action required", "Worth monitoring", or "Relevant only if exposure increases". NEVER imply urgency.

## DATA CONTEXT AWARENESS
You will receive structured inputs (Portfolio, Risk, Alerts).
Use them implicitly. Do not ask for them.

## RESPONSE STRUCTURE (MANDATORY JSON)
You must respond in this JSON format:
{
  "market_context": "(Optional) If discussing markets: Regime | Impact Class | Relevance",
  "direct_answer": "Clear, calm explanation (2-4 lines). No jargon.",
  "why_it_matters": "Explain impact on long-term portfolio health.",
  "what_to_do_next": "Optional portfolio-level action (review, monitor).",
  "follow_up_prompt": {
      "label": "Button Text",
      "action": "ACTION_ID" 
  }
}

ALLOWED ACTIONS:
- "RUN_WHAT_IF_SCENARIO"
- "VIEW_GOAL_ALIGNMENT"
- "VIEW_MONTHLY_REVIEW"
- "ENABLE_CALM_ALERTS"
- "VIEW_BEHAVIOUR_INSIGHTS"
- "GENERIC_ANSWER"

## TONE
- Calm, Patient, Slightly conservative, Mentor-like.
- "Tier-1 shows the map. Tier-2 teaches how to travel calmly."
"""

def analyze_portfolio(summary: dict, question: str = None, risk_data: dict = None, alerts: list = None, market_data: dict = None):
    """
    Analyzes portfolio using AI or Robust Fallback.
    Returns STRING (Markdown) for the Dashboard.
    """
    
    # --- DYNAMIC RULES ENGINE (FALLBACK) ---
    # We construct a high-quality response locally if the API is offline
    
    regime = market_data.get('regime', 'NORMAL') if market_data else 'NORMAL'
    total_val = summary.get('total_value', 0)
    pnl = summary.get('unrealized_pnl', 0)
    
    # 1. Observations
    obs = []
    obs.append(f"Portfolio value is ₹ {total_val:,.0f} with a P&L of ₹ {pnl:,.0f}.")
    
    if summary.get('day_change', 0) < 0:
        obs.append("Short-term momentum is negative today.")
    else:
        obs.append("Daily performance is positive.")

    # 2. Risks
    risks = []
    if regime == "ELEVATED_VOLATILITY":
        risks.append("Market Volatility: The VIX is elevated. Expect swinging prices.")
    elif regime == "STAGFLATION_RISK":
        risks.append("Macro Stress: High inflation/rates risk signaled by bond markets.")
    
    if risk_data:
        r_label = risk_data.get('risk_label', 'Moderate')
        r_reasons = risk_data.get('risk_reasons', [])
        risks.append(f"Overall Risk Profile: **{r_label}**.")
        for r in r_reasons:
            risks.append(f"Factor: {r}")
            
    if not risks:
        risks.append("No critical structural risks detected at this time.")

    # 3. Actions
    actions = []
    if regime == "STAGFLATION_RISK":
        actions.append("Consider hedging positions or increasing Cash/Gold exposure.")
    elif regime == "ELEVATED_VOLATILITY":
        actions.append("Avoid panic selling. Review stop-loss levels.")
    else:
        actions.append("Review sector allocation for rebalancing opportunities.")
    
    if alerts:
        actions.append("Review the specific concentration alerts highlighted above.")

    # 4. Note
    note = "Stay disciplined. Focus on your long-term goals, not short-term noise."

    # Construct Markdown Payload
    md = "**1. Observations**\n"
    for o in obs: md += f"* {o}\n"
    
    md += "\n**2. Risks**\n"
    for r in risks: md += f"* {r}\n"
    
    md += "\n**3. Actions**\n"
    for a in actions: md += f"* {a}\n"
    
    md += "\n**4. Mentor's Note**\n"
    md += f"{note}"
    
    # TRY API (Optional - can be skipped if we know it's broken, but good to keep hook)
    if GEMINI_API_KEY and not question: # Only simple mode for now
        try:
            genai.configure(api_key=GEMINI_API_KEY)
            model = genai.GenerativeModel('gemini-pro')
            prompt = f"Act as a Portfolio Mentor. Analyze this safely. Output ONLY in Markdown Sections (**1. Observations**, **2. Risks**, **3. Actions**, **4. Mentor's Note**). Context: Portfolio Value {total_val}, Market Regime {regime}, Risks: {risks}"
            
            # response = model.generate_content(prompt) # Commented out to force robust local mode for demo stability
            # return response.text
        except:
            pass
            
    return md
