requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
apscheduler
google-generativeai
yfinance
pandas
numpy
